requirements.txt
# Core Dependencies
streamlit==1.31.0
python-dotenv==1.0.0

# AI/ML Dependencies
huggingface-hub==0.20.3
transformers==4.37.2
torch==2.1.2
accelerate==0.26.1
bitsandbytes==0.42.0

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Visualization
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.1

# API & Utils
requests==2.31.0
Pillow==10.2.0

# Optional: For better performance
sentencepiece==0.1.99
protobuf==4.25.2